fix relaying of signals, child now stays alive if it exits by itself

+++ hickup.js
@@ -3,34 +3,66 @@
 import { spawn } from "child_process"
 
 const killTimeout = 10000
+const relayedSignals = ["SIGILL", "SIGTRAP", "SIGABRT", "SIGIOT", "SIGBUS", "SIGFPE", "SIGUSR1", "SIGSEGV", "SIGUSR2", "SIGPIPE", "SIGALRM", "SIGCHLD", "SIGSTKFLT", "SIGCONT", "SIGTSTP", "SIGBREAK", "SIGTTIN", "SIGTTOU", "SIGURG", "SIGXCPU", "SIGXFSZ", "SIGVTALRM", "SIGPROF", "SIGWINCH", "SIGIO", "SIGPOLL", "SIGLOST", "SIGPWR", "SIGSYS", "SIGUNUSED"]
 
-function spawner() {
-  if (process.argv.length < 3) {
-    process.exit(1)
-  }
-  return spawn(process.argv[2], process.argv.slice(3), { stdio: "inherit" })
+let child, timeout, alive, killed
+
+if (process.argv.length < 3) {
+  process.exit(1)
 }
 
-let child = spawner()
+process
+  .on("SIGHUP", respawnChild)
+  .on("SIGTERM", killSelf)
+  .on("SIGINT", killSelf)
+  .on("SIGQUIT", killSelf)
 
-process.on("SIGHUP", () => {
-  console.log("SIGHUP: Restarting...")
+relayedSignals.forEach(signal => process.on(signal, () => child.kill(signal)))
+
+function killSelf() {
+  killChild()
+  process.exit(0)
+}
+
+function killChild() {
+  killed = true
   child.kill()
+  if (!timeout) {
+    timeout = setTimeout(() => child.kill("SIGKILL"), killTimeout)
+  }
+}
 
-  let killed = false
-  const timeout = setTimeout(() => child.kill("SIGKILL"), killTimeout)
+function spawnChild() {
+  if (!alive) {
+    child = spawn(process.argv[2], process.argv.slice(3), { stdio: "inherit" })
+      .once("error", cleanUp)
+      .once("exit", cleanUp)
+    alive = true
+    killed = false
+  }
+}
 
-  function cleanUp() {
-    if (!killed) {
-      killed = true
-      clearTimeout(timeout)
-      child = spawner()
+function respawnChild() {
+  if (alive) {
+    killChild()
+  } else {
+    spawnChild()
+  }
+}
+
+function cleanUp() {
+  if (alive) {
+    alive = false
+    if (killed) {
+      respawnChild()
+      if (timeout) {
+        clearTimeout(timeout)
+        timeout = null
+      }
+    } else {
+      process.stdin.resume()
     }
   }
+}
 
-  child
-  .once("error", cleanUp)
-  .once("exit", cleanUp)
-})
-
-// TODO: signal propagation
+spawnChild()
